Code cleanup

+++ collect-core/src/main/java/org/openforis/collect/manager/SessionRecordFileManager.java
@@ -47,6 +47,7 @@ public class SessionRecordFileManager implements Serializable {
 	
 	public void destroy() {
 		deleteAllTempFiles();
+		resetTempInfo();
 	}
 
 	public void resetTempInfo() {
@@ -120,7 +121,13 @@ public class SessionRecordFileManager implements Serializable {
 				filesToDelete.put(nodeId, fileName);
 			}
 		} else {
-			//remove temp file
+			deleteTempFile(record, nodeId);
+		}
+	}
+	
+	public void deleteTempFile(CollectRecord record, int nodeId) {
+		TempFileInfo tempFileInfo = nodeIdToTempFilePath.get(nodeId);
+		if (tempFileInfo != null ) {
 			java.io.File tempFile = new java.io.File(tempFileInfo.getTempFilePath());
 			tempFile.delete();
 			nodeIdToTempFilePath.remove(nodeId);
